requirements.txt
streamlit
matplotlib
numpy
pandas
st-circular-progress
Pillow
